Lint-only errors return exit status 0
Updating code to examine the type of errors a bit more closely,
specifically if there are only linting errors. The reason for this is
that we don't want to fail external build tools if there are linting
errors. We are setting the exit status code based on the type of error(s).